Created a button that takes you to the next screen
Implemented a navigation hierarchy and organized my components into a folder
