+++ App.jsx
@@ -1,0 +1,25 @@
+import { StyleSheet} from 'react-native';
+import {NavigationContainer} from '@react-navigation/native';
+import {createStackNavigator} from '@react-navigation/stack';
+import WelcomeScreen from './components/WelcomeScreen';
+
+
+
+export default function App() {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator intialRouteName="WelcomeScreen">
+        <Stack.Screen name="WelcomeScreen" component={WelcomeScreen}/>
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
