+++ components/WelcomeScreen.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { View } from 'react-native'
+import NextScreenBtn from './NextScreenBtn'
+
+const WelcomeScreen = () => {
+  return (
+    <View>
+      <NextScreenBtn btnText="Create New Bill"/>
+    </View>
+  )
+}
+
+export default WelcomeScreen
